flashcard_backend/src/users/users.controller.ts
import { Controller, Post, Body, ConflictException, InternalServerErrorException } from '@nestjs/common';
import { UsersService } from './users.service';
import { CreateUserDto } from './dto/user.dto';

@Controller('users')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @Post('register')
  async register(@Body() body: CreateUserDto ) {
    try {
      const user = await this.usersService.create(body);
      return { message: 'User registered successfully', userId: user.id };
    } catch (error) {
      if (error instanceof ConflictException) {
        throw error;
      }
      throw new InternalServerErrorException('Error registering user');
    }
  }
}
